+++ packages/prettier/src/options/sort-vue.ts
@@ -4,7 +4,7 @@ import organizeAttributes from '../plugin-organize-attributes'
 // https://prettier.io/docs/options
 // https://github.com/NiklasPor/prettier-plugin-organize-attributes
 export default {
-  parser: 'html',
+  parser: 'vue',
   plugins: [organizeAttributes],
   attributeSort: 'ASC',
   attributeIgnoreCase: true,